refactor: 대시보드 회의 아이템 공용 컴포넌트로 분리 #82
