+++ src/features/owner/lib.test.ts
@@ -1,4 +1,4 @@
-import { formatMeetingDate, getDaysUntilDue } from "./lib";
+import { getDaysUntilDue } from "./lib";
 
 describe("getDaysUntilDue", () => {
   beforeEach(() => {
@@ -19,15 +19,5 @@ describe("getDaysUntilDue", () => {
 
   it("지난 날짜는 음수를 반환한다", () => {
     expect(getDaysUntilDue("2026-08-01")).toBe(-4);
-  });
-});
-
-describe("formatMeetingDate", () => {
-  it("월 일(요일) 시:분 형식으로 포맷한다", () => {
-    expect(formatMeetingDate("2026-08-20T10:00:00")).toBe("8월 20일(목) 10:00");
-  });
-
-  it("한 자리 시·분도 두 자리로 채운다", () => {
-    expect(formatMeetingDate("2026-09-01T09:05:00")).toBe("9월 1일(화) 09:05");
   });
 });
