+++ src/features/owner/mock/dashboard.ts
@@ -80,6 +80,7 @@ export const OWNER_DASHBOARD_MOCK: OwnerDashboardOverview = {
       room: "회의실 A",
       scheduledAt: "2026-07-30T10:00:00",
       attendeeCount: 4,
+      hostLabel: "Owner",
     },
     {
       id: "m2",
@@ -90,6 +91,7 @@ export const OWNER_DASHBOARD_MOCK: OwnerDashboardOverview = {
       room: "대회의실",
       scheduledAt: "2026-08-01T14:00:00",
       attendeeCount: 4,
+      hostLabel: "Owner",
     },
     {
       id: "m3",
@@ -100,6 +102,7 @@ export const OWNER_DASHBOARD_MOCK: OwnerDashboardOverview = {
       room: "회의실 B",
       scheduledAt: "2026-08-05T11:00:00",
       attendeeCount: 3,
+      hostLabel: "Owner",
     },
     {
       id: "m4",
@@ -110,6 +113,7 @@ export const OWNER_DASHBOARD_MOCK: OwnerDashboardOverview = {
       room: "회의실 C",
       scheduledAt: "2026-08-08T15:00:00",
       attendeeCount: 3,
+      hostLabel: "Owner",
     },
     {
       id: "m5",
@@ -120,6 +124,7 @@ export const OWNER_DASHBOARD_MOCK: OwnerDashboardOverview = {
       room: "회의실 A",
       scheduledAt: "2026-08-12T10:00:00",
       attendeeCount: 2,
+      hostLabel: "Owner",
     },
   ],
 };
